+++ MoreShipUpgrades/Managers/UpgradeBus.cs
@@ -1,6 +1,4 @@
 ﻿using GameNetcodeStuff;
-using LethalLib.Extras;
-using LethalLib.Modules;
 using MoreShipUpgrades.Configuration;
 using MoreShipUpgrades.Configuration.Upgrades.Interfaces;
 using MoreShipUpgrades.Configuration.Upgrades.Interfaces.OneTimeUpgrades;
@@ -42,8 +40,6 @@ namespace MoreShipUpgrades.Managers
         internal AudioClip flashNoise;
         internal GameObject modStorePrefab;
 
-        internal Dictionary<string, SpawnableMapObjectDef> spawnableMapObjects = [];
-        internal Dictionary<string, int> spawnableMapObjectsAmount = [];
         internal readonly List<Type> upgradeTypes = [];
         internal readonly List<Type> commandTypes = [];
         internal readonly List<Type> itemTypes = [];
@@ -176,14 +172,14 @@ namespace MoreShipUpgrades.Managers
             Item storeItem = ItemsToSync[itemName];
             if (!configuredEnable)
             {
-                Items.RemoveShopItem(storeItem);
+                ItemManager.RemoveShopItem(storeItem);
                 logger.LogInfo($"Removing {itemName} from store.");
                 return;
             }
             if (storeItem.creditsWorth != configuredPrice)
             {
                 logger.LogInfo($"Changing {itemName}'s price from {storeItem.creditsWorth} to {configuredPrice}");
-                Items.UpdateShopItemPrice(storeItem, configuredPrice);
+                ItemManager.UpdateShopItemPrice(storeItem, configuredPrice);
             }
         }
 
@@ -195,7 +191,7 @@ namespace MoreShipUpgrades.Managers
 
         void SyncAvailableContracts()
         {
-            if (!PluginConfiguration.DATA_CONTRACT.Value || !PluginConfiguration.CONTRACTS_ENABLED.Value)
+            if (!PluginConfiguration.ContractsConfiguration.DataConfiguration.Enabled.Value || !PluginConfiguration.ContractsConfiguration.Enabled.Value)
             {
                 logger.LogInfo("Removing data contract");
                 int idx = CommandParser.contracts.IndexOf(LguConstants.DATA_CONTRACT_NAME);
@@ -205,7 +201,7 @@ namespace MoreShipUpgrades.Managers
                     CommandParser.contracts.RemoveAt(idx);
                 }
             }
-            if (!PluginConfiguration.EXTRACTION_CONTRACT.Value || !PluginConfiguration.CONTRACTS_ENABLED.Value)
+            if (!PluginConfiguration.ContractsConfiguration.ExtractionConfiguration.Enabled.Value || !PluginConfiguration.ContractsConfiguration.Enabled.Value)
             {
                 logger.LogInfo("Removing extraction contract");
                 int idx = CommandParser.contracts.IndexOf(LguConstants.EXTRACTION_CONTRACT_NAME);
@@ -215,7 +211,7 @@ namespace MoreShipUpgrades.Managers
                     CommandParser.contracts.RemoveAt(idx);
                 }
             }
-            if (!PluginConfiguration.EXORCISM_CONTRACT.Value || !PluginConfiguration.CONTRACTS_ENABLED.Value)
+            if (!PluginConfiguration.ContractsConfiguration.ExorcismConfiguration.Enabled.Value || !PluginConfiguration.ContractsConfiguration.Enabled.Value)
             {
                 logger.LogInfo("Removing exorcism contract");
                 int idx = CommandParser.contracts.IndexOf(LguConstants.EXORCISM_CONTRACT_NAME);
@@ -225,7 +221,7 @@ namespace MoreShipUpgrades.Managers
                     CommandParser.contracts.RemoveAt(idx);
                 }
             }
-            if (!PluginConfiguration.DEFUSAL_CONTRACT.Value || !PluginConfiguration.CONTRACTS_ENABLED.Value)
+            if (!PluginConfiguration.ContractsConfiguration.DefusalConfiguration.Enabled.Value || !PluginConfiguration.ContractsConfiguration.Enabled.Value)
             {
                 logger.LogInfo("Removing defusal contract");
                 int idx = CommandParser.contracts.IndexOf(LguConstants.DEFUSAL_CONTRACT_NAME);
@@ -235,9 +231,9 @@ namespace MoreShipUpgrades.Managers
                     CommandParser.contracts.RemoveAt(idx);
                 }
             }
-            if (!PluginConfiguration.EXTERMINATOR_CONTRACT.Value || !PluginConfiguration.CONTRACTS_ENABLED.Value)
+            if (!PluginConfiguration.ContractsConfiguration.ExterminationConfiguration.Enabled.Value || !PluginConfiguration.ContractsConfiguration.Enabled.Value)
             {
-                if(CommandParser.contracts.Count == 1 && PluginConfiguration.CONTRACTS_ENABLED.Value)
+                if(CommandParser.contracts.Count == 1 && PluginConfiguration.ContractsConfiguration.Enabled.Value)
                 {
                     logger.LogInfo("Why must you do the things you do");
                     logger.LogWarning("User tried to remove all contracts! Leaving exterminator present. Did you mean to disable contracts?");