+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Ship/FusionMatter.cs
@@ -150,9 +150,9 @@ namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.Ship
         }
         public new static void RegisterUpgrade()
         {
-            GameObject prefab = LethalLib.Modules.NetworkPrefabs.CreateNetworkPrefab(UPGRADE_NAME);
+            GameObject prefab = ItemManager.CreateNetworkPrefab(UPGRADE_NAME);
             prefab.AddComponent<FusionMatter>();
-            LethalLib.Modules.NetworkPrefabs.RegisterNetworkPrefab(prefab);
+            ItemManager.RegisterNetworkPrefab(prefab);
             Plugin.networkPrefabs[UPGRADE_NAME] = prefab;
         }
         public new static CustomTerminalNode RegisterTerminalNode()